+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Breaking changes
 - x/paychan is using schema versioned model and messages
 - x/escrow is using schema versioned model and messages
 - x/distribution is using schema versioned model and messages
+- x/namecoin is using schema versioned model and messages
+- x/sigs is using schema versioned model and messages
 
 
 ## 0.13.0